Update: :sparkles: package.json have dev and build script
On the dev mode. You could write CLI gulp dev and on the production. You write gulp --prod, but now instead we write npm run dev or npm run build instead. This is because we don't always remember adding a --flag which is gulp --build. The usually convention way is to write npm run <whatever-here> to remember to run the automation development workflow instead. It's easier to remember