Add: :sparkles: jQuery to import WP Core
The jQuery is already in-built for WordPress Core. You should never download an dependency from jQuery library of whatever version. This is because WP has their own default jQuery version installed on the WP Core. If you try to modify and use different jQuery version. The theme, plugins whatever you are trying to use might break the website