Add: :sparkles: Browsersync and CSS series injection
This commit will ensure that we don't need to manually reload the website when we are doing changes. Using the Browsersync will automatically take care it for us.

The .pipe(server.stream()) on arrow function of styles will make sure we
don't need to wait to refresh and reload the website each time we are
doing changes on the SCSS. It pretty straight injecting the CSS in the
browser document without even refreshing the page.

The proxy is using a "mini" server in order to use the Browsersync. We
have to make sure that URL from the proxy needs to be exactly like from
your URL address when you are running the local wordpress on your local
machine. The url could be for example http://localhost:8888/<theme-name>
if you are using LAMP/WAMP/XAMPP. Since we are using the Local WP, the
URL address will be http://dummy-theme.local/ which actually translates
the http://<Wordpress-site-name>.local/