Add: :sparkles: gulp-replace to replace WP prefix placeholder
This is used to replace the filename of every prefix that exist on our code source. The prefix is _theme_name, so by using the gulp-replace, we will replace the prefix from _theme_name to dummy_theme.The gulp-replace will make sure to find every name that exist on the file we have in our project that has the _theme_name and replace it for us. This will make sure that it's very easy for us in the future to make a new theme and reuse some code from our old themes.

The dependency link: https://www.npmjs.com/package/gulp-replace