Add: :sparkles: Created a demo custom action hooks
Usually in WP Core, there are plenty of different hooks that we can use,
such as action and filters:
https://developer.wordpres.org/plugins/hooks/actions/
https://developer.wordpress.org/plugins/hooks/
https://codex.wordpress.org/Plugin_API/Filter_Reference
https://codex.wordpress.org/Plugin_API/Action_Reference

The difference with the do_action() is that we can create our own
customized hooks. That hooks can be used for current project, and if
there are other developers who wants to use our hooks, they can do it by
using the add_action('your-customized-hooks', cb);

This customized hooks in this commit doesn't do anything special than
just echoing out strings. It's only for demo purpose