Add: :sparkles: add_action and add_filter as a demo
add_action() is impure function. That means we can expect to have side
effect

In this example, we have used a hook that is called 'pre_get_posts' and the function used to register an action get passed the current instance of the WP_query, and then we use the is_main_query() from the WP_query global instance  which exist in the WP core. By that, we can filter anything, and in this case, we just want to
filter by getting only 2 posts before the global wp_query is being
triggered.

add_filter() is pure function. We expect:

1. You get an input
2. You modify it
3. You return it

Example in this case we modify the text of add_filter. We expect we only
get the change that is text based which is a string. Nothing else!