Add: :sparkles: Organized the post template for DRY
the loop.php is make, or rather making our WP files to be better structured. We want to keep it DRY.

We could also use include or requirece php function, but we don't want
to do it. instead, we use get_tempalte_part() from the WP. The
difference is that WP will look for the file which in this commit is
loop.php in our child theme. If it's not found, then it will fall back
to the parent theme.

Just imagine some other developer wants to modify something in your
loop. Now, that developer will only have to create a file called
loop.php in his child theme instead of creating the whole archive page
and copying this whole lines of code from the original archive.php file
just to modify something in the loop.