Add: :sparkles: Theme support for header
Added the title-tag, post-thumbnails and html5.

html5 is to tell WP explicitly choose to apply HTML5 markup for search forms, comment forms, comment lists, gallery and caption. This is important, because we are telling the for example get_search_form() will generate the html for the search form, but before that, it would actually search your theme folder is called searchform.php