Add: :sparkles: PostMessage preview in-live when editing footer
Before when changing on the theme customization. You usually have to wait to reload automatically. That is because WP is using iframe to communicate between the wp database and tthe client. By using this technique of transport. We will able to display information right away, because JavaScript will help us with it. It's like onInput you could say.

You can look at the chart and more information here:
https://developer.wordpress.org/files/2017/01/Customize-Object-Hierarchy-Graphic-768x432.png[201~

https://developer.wordpress.org/themes/customize-api/customizer-objects/