Add: :sparkles: Selective refresh for iframe customization
Previous commits, the WP will refresh the whole iframe when doing
real-live change on the Theme customization. In this commit, instead of
refreshing the whole website. We just choose which part of the
cuztomization we want to the WP iframe to refresh specific input section. In this case, it's the site title