Add: :sparkles: Added custom logo for WP Customization
Now you can add your own custom logo to display on the front-end. There is an issue with WP though. When removing the logo. The default text doesn't show right away. The key is to just type something on the text field so WP can refresh the iframe of the site in order to make it work again