Add: :art: Improved inline styles selector
If we have tons of CSS selectors and properties. This commit will be
much better way to write cleaner and seperate the logic of CSS on the
inline-css.php. The logic is that PHP must be looped through first of
the logic, and then JavaScript will mimick after the PHP loops