Add: :sparkles: non-nonce users cannot delete post from URL address
This is a security for making sure if someone got our URL link by accident, malware etc. They cannot delete our WP post because we use nonce technique that WP provides. The user needs to login and have the correct type or user role in order to delete.

To understand a bit more what nonce do:
1. nonce will create hash and the hash will contain information like
   current user ID and current user sesttion token
2. Session token is generated when logging into WP and it's unique!
3. If you login from another browser, but same account. It will create
   another unique token
4. Nonce also generate 24 hours time frame. After 24 hours. It will
   delete the token