Add: ✨  A New Gulp Task To Copy Plugins
When we have created a production plugin file from our plugin/ folder
of dummytheme-metaboxes.
Next step is using our dummy-theme to make a zip file of theme production
theme. When we do that, the production zip theme will make sure to
take the copy of the production dummy metaboxes plugin and insert it to
our production theme. This will ensure better workflow to be faster
than we have to manually copy the production plugin inside our
theme folder of dummy/lib/plugins/<plugin-file>